requirements.txt
Django
boto
dj-database-url
django-storages
gunicorn
psycopg2
wsgiref
south
django-disqus 
django-menu 
django-taggit 
django-any
django-jenkins
django-tinymce
django-filebrowser==3.5.3
Pillow